+++ packages/fetch-router/src/index.ts
@@ -13,6 +13,23 @@ export type { RequestHandler, BuildRequestHandler } from './lib/request-handler.
 export type { Controller, Action, BuildAction } from './lib/controller.ts'
 
 export {
+  createDestroy,
+  createDestroy as destroy, // shorthand
+  createGet,
+  createGet as get, // shorthand
+  createHead,
+  createHead as head, // shorthand
+  createOptions,
+  createOptions as options, // shorthand
+  createPatch,
+  createPatch as patch, // shorthand
+  createPost,
+  createPost as post, // shorthand
+  createPut,
+  createPut as put, // shorthand
+} from './lib/route-helpers.ts'
+
+export {
   Route,
   createRoutes,
   createRoutes as route, // shorthand